improve(@bud) read/write config values with api

+++ dev/@bud/sources/command/clean.command.ts
@@ -2,11 +2,26 @@ import {CommandClass, Option} from 'clipanion'
 
 import {Command} from './base.command'
 
+/**
+ * Clean command
+ *
+ * @internal
+ */
 export class Clean extends Command {
+  /**
+   * Command paths
+   *
+   * @internal
+   */
   public static paths: CommandClass['paths'] = [
     [`@bud`, `clean`],
   ]
 
+  /**
+   * Command usage
+   *
+   * @internal
+   */
   public static usage: CommandClass['usage'] = {
     category: `@bud`,
     description: `clean project artifacts`,
@@ -15,8 +30,22 @@ export class Clean extends Command {
     ],
   }
 
+  /**
+   * --all flag
+   *
+   * @internal
+   */
+  public all = Option.Boolean(`-a,--all`, false, {
+    description: `cleans all untracked files`,
+  })
+
+  /**
+   * Command execution
+   *
+   * @internal
+   */
   public async execute() {
     await this.$(`yarn cache clean --all`)
-    await this.$(`git clean -fxd`)
+    if (this.all) await this.$(`git clean -fxd`)
   }
 }
